move percent data to `percent.rs`
