+++ components/experimental/src/dimension/provider/mod.rs
@@ -3,3 +3,4 @@
 // (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
 
 pub mod currency;
+pub mod percent;
