+++ provider/datagen/tests/data/baked/macros/percent_essentials_v1.rs.data
@@ -9,15 +9,15 @@ macro_rules! __impl_percent_essentials_v1 {
         #[clippy::msrv = "1.67"]
         const _: () = <$provider>::MUST_USE_MAKE_PROVIDER_MACRO;
         #[clippy::msrv = "1.67"]
-        impl icu_provider::DataProvider<icu_experimental::dimension::provider::currency::PercentEssentialsV1Marker> for $provider {
-            fn load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponse<icu_experimental::dimension::provider::currency::PercentEssentialsV1Marker>, icu_provider::DataError> {
-                static UND: <icu_experimental::dimension::provider::currency::PercentEssentialsV1Marker as icu_provider::DataMarker>::Yokeable = icu_experimental::dimension::provider::currency::PercentEssentialsV1 { number_index: 0u8, percent_sign_affixes: icu_experimental::dimension::provider::currency::PercentAffixesV1 { prefix: alloc::borrow::Cow::Borrowed(""), suffix: alloc::borrow::Cow::Borrowed("") }, percent_sign_symbol: alloc::borrow::Cow::Borrowed("%"), percent_symbol_index: 5u8, standard: alloc::borrow::Cow::Borrowed("#,##0%") };
-                static VALUES: [&<icu_experimental::dimension::provider::currency::PercentEssentialsV1Marker as icu_provider::DataMarker>::Yokeable; 1usize] = [&UND];
+        impl icu_provider::DataProvider<icu_experimental::dimension::provider::percent::PercentEssentialsV1Marker> for $provider {
+            fn load(&self, req: icu_provider::DataRequest) -> Result<icu_provider::DataResponse<icu_experimental::dimension::provider::percent::PercentEssentialsV1Marker>, icu_provider::DataError> {
+                static UND: <icu_experimental::dimension::provider::percent::PercentEssentialsV1Marker as icu_provider::DataMarker>::Yokeable = icu_experimental::dimension::provider::percent::PercentEssentialsV1 { number_index: 0u8, percent_sign_affixes: icu_experimental::dimension::provider::percent::PercentAffixesV1 { prefix: alloc::borrow::Cow::Borrowed(""), suffix: alloc::borrow::Cow::Borrowed("") }, percent_sign_symbol: alloc::borrow::Cow::Borrowed("%"), percent_symbol_index: 5u8, standard: alloc::borrow::Cow::Borrowed("#,##0%") };
+                static VALUES: [&<icu_experimental::dimension::provider::percent::PercentEssentialsV1Marker as icu_provider::DataMarker>::Yokeable; 1usize] = [&UND];
                 static KEYS: [&str; 1usize] = ["und"];
                 if let Ok(payload) = KEYS.binary_search_by(|k| req.locale.strict_cmp(k.as_bytes()).reverse()).map(|i| *unsafe { VALUES.get_unchecked(i) }) {
                     Ok(icu_provider::DataResponse { payload: Some(icu_provider::DataPayload::from_static_ref(payload)), metadata: Default::default() })
                 } else {
-                    Err(icu_provider::DataErrorKind::MissingLocale.with_req(<icu_experimental::dimension::provider::currency::PercentEssentialsV1Marker as icu_provider::KeyedDataMarker>::KEY, req))
+                    Err(icu_provider::DataErrorKind::MissingLocale.with_req(<icu_experimental::dimension::provider::percent::PercentEssentialsV1Marker as icu_provider::KeyedDataMarker>::KEY, req))
                 }
             }
         }
